+++ multidid/src/lib.rs
@@ -5,17 +5,13 @@ use unsigned_varint::{
     io::read_u64,
 };
 
-const RAW_CODEC: u64 = 0x55;
-const PKH_CODEC: u64 = 0xca;
-const SECP256K1_CODEC: u64 = 0xe7;
-const BLS12_381_G1_CODEC: u64 = 0xea;
-const BLS12_381_G2_CODEC: u64 = 0xeb;
-const X25519_CODEC: u64 = 0xec;
-const ED25519_CODEC: u64 = 0xed;
-const P256_CODEC: u64 = 0x1200;
-const P384_CODEC: u64 = 0x1201;
-const P521_CODEC: u64 = 0x1202;
-// const RSA_CODEC: u64 = 0x1205;
+mod error;
+mod key;
+mod method;
+
+pub use error::Error;
+pub use key::DidKeyTypes;
+pub use method::Method;
 
 const MULTIDID_VARINT_TAG: u16 = 0x9d1a;
 
@@ -103,7 +99,7 @@ impl MultiDid {
         })
     }
 
-    pub fn to_writer<W: Write>(&self, writer: &mut W) -> Result<(), Error> {
+    pub fn to_writer<W: Write>(&self, writer: &mut W) -> Result<(), IoError> {
         writer.write_all(&MULTIDID_VARINT_TAG.to_be_bytes())?;
         // write codec
         let mut buf = u64_buffer();
@@ -158,175 +154,6 @@ impl MultiDid {
             (None, None) => {}
         };
         Ok(())
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Method {
-    Pkh(Vec<u8>),
-    Key(DidKeyTypes),
-    Raw(String),
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error(transparent)]
-    Varint(#[from] unsigned_varint::io::ReadError),
-    #[error(transparent)]
-    Io(#[from] IoError),
-    #[error("Invalid multidid varint prefix, expected 0x9d1a, recieved {0:x}")]
-    InvalidPrefix(u64),
-    #[error(transparent)]
-    Parameter(#[from] iri_string::validate::Error),
-    #[error(transparent)]
-    DidParse(#[from] std::string::FromUtf8Error),
-}
-
-impl Method {
-    pub fn codec(&self) -> u64 {
-        use Method::*;
-        match self {
-            Raw(_) => RAW_CODEC,
-            Pkh(_) => PKH_CODEC,
-            Key(k) => k.codec(),
-        }
-    }
-
-    fn from_reader<R: Read>(reader: &mut R) -> Result<Self, Error> {
-        let codec = read_u64(reader.by_ref())?;
-        match codec {
-            RAW_CODEC => {
-                let len = read_u64(reader.by_ref())?;
-                let mut buf = vec![0; len as usize];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Raw(String::from_utf8(buf)?))
-            }
-            PKH_CODEC => {
-                let len = read_u64(reader.by_ref())?;
-                let mut buf = vec![0; len as usize];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Pkh(buf))
-            }
-            codec => {
-                let key = DidKeyTypes::from_reader(reader, codec)?;
-                Ok(Self::Key(key))
-            }
-        }
-    }
-
-    fn to_writer<W: Write>(&self, writer: &mut W) -> Result<(), IoError> {
-        match self {
-            Self::Raw(buf) => {
-                writer.write_all(buf.as_bytes())?;
-            }
-            Self::Pkh(buf) => {
-                let mut vi_buf = u64_buffer();
-                writer.write_all(write_u64(buf.len() as u64, &mut vi_buf))?;
-                writer.write_all(buf)?;
-            }
-            Self::Key(key) => {
-                key.to_writer(writer)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum DidKeyTypes {
-    Secp256k1([u8; 33]),
-    Bls12_381G1([u8; 64]),
-    Bls12_381G2([u8; 96]),
-    X25519([u8; 32]),
-    Ed25519([u8; 32]),
-    P256([u8; 33]),
-    P384([u8; 49]),
-    P521([u8; 67]),
-    // RSA([u8; ??]),
-}
-
-impl DidKeyTypes {
-    pub fn codec(&self) -> u64 {
-        use DidKeyTypes::*;
-        match self {
-            Secp256k1(_) => SECP256K1_CODEC,
-            Bls12_381G1(_) => BLS12_381_G1_CODEC,
-            Bls12_381G2(_) => BLS12_381_G2_CODEC,
-            X25519(_) => X25519_CODEC,
-            Ed25519(_) => ED25519_CODEC,
-            P256(_) => P256_CODEC,
-            P384(_) => P384_CODEC,
-            P521(_) => P521_CODEC,
-        }
-    }
-
-    fn from_reader<R>(reader: &mut R, codec: u64) -> Result<Self, Error>
-    where
-        R: Read,
-    {
-        match codec {
-            SECP256K1_CODEC => {
-                let mut buf = [0; 33];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Secp256k1(buf))
-            }
-            BLS12_381_G1_CODEC => {
-                let mut buf = [0; 64];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Bls12_381G1(buf))
-            }
-            BLS12_381_G2_CODEC => {
-                let mut buf = [0; 96];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Bls12_381G2(buf))
-            }
-            X25519_CODEC => {
-                let mut buf = [0; 32];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::X25519(buf))
-            }
-            ED25519_CODEC => {
-                let mut buf = [0; 32];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::Ed25519(buf))
-            }
-            P256_CODEC => {
-                let mut buf = [0; 33];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::P256(buf))
-            }
-            P384_CODEC => {
-                let mut buf = [0; 49];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::P384(buf))
-            }
-            P521_CODEC => {
-                let mut buf = [0; 67];
-                reader.read_exact(&mut buf)?;
-                Ok(Self::P521(buf))
-            }
-            _ => Err(Error::InvalidPrefix(codec)),
-        }
-    }
-
-    fn to_writer<W>(&self, writer: &mut W) -> Result<(), IoError>
-    where
-        W: ?Sized + Write,
-    {
-        writer.write_all(&self.bytes())
-    }
-
-    fn bytes(&self) -> &[u8] {
-        match self {
-            Self::Secp256k1(key) => key,
-            Self::Bls12_381G1(key) => key,
-            Self::Bls12_381G2(key) => key,
-            Self::X25519(key) => key,
-            Self::Ed25519(key) => key,
-            Self::P256(key) => key,
-            Self::P384(key) => key,
-            Self::P521(key) => key,
-        }
     }
 }
